fix: split test code
